# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Place {
  id: number;
  tripId: number;
  name: string;
  city: string;
  category: string;
  description: string;
  imageUrl: string;
  voteCount: number;
  isVoted: boolean;
}

export interface Trip {
  id: number;
  title: string;
  destination: string;
  dates: string;
  description: string;
  coverImageUrl: string;
  placeCount: number;
  totalVotes: number;
  places: Place[];
}

export interface TripSummary {
  id: number;
  title: string;
  destination: string;
  dates: string;
  description: string;
  coverImageUrl: string;
  placeCount: number;
  totalVotes: number;
  /** @nullable */
  leadingPlace: string | null;
}

export interface PlaceVoteSummary {
  placeId: number;
  name: string;
  voteCount: number;
}

export interface TripVoteSummary {
  tripId: number;
  totalVotes: number;
  /** @nullable */
  leadingPlace: string | null;
  places: PlaceVoteSummary[];
}

export type VoteInputMode = typeof VoteInputMode[keyof typeof VoteInputMode];


export const VoteInputMode = {
  solo: 'solo',
  group: 'group',
} as const;

export interface VoteInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  voterId: string;
  mode: VoteInputMode;
}

export type VoteResultMode = typeof VoteResultMode[keyof typeof VoteResultMode];


export const VoteResultMode = {
  solo: 'solo',
  group: 'group',
} as const;

export interface VoteResult {
  tripId: number;
  placeId: number;
  voteCount: number;
  totalVotes: number;
  isVoted: boolean;
  mode: VoteResultMode;
}

export type VoterIdParameter = string;

export type GetTripParams = {
/**
 * @maxLength 128
 */
voterId?: VoterIdParameter;
};

export type GetTripSummaryParams = {
/**
 * @maxLength 128
 */
voterId?: VoterIdParameter;
};