# lib/api-zod/src/generated/types/place.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Place {
  id: number;
  tripId: number;
  name: string;
  city: string;
  category: string;
  description: string;
  imageUrl: string;
  voteCount: number;
  isVoted: boolean;
}