# lib/api-zod/src/generated/types/trip.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Place } from './place';

export interface Trip {
  id: number;
  title: string;
  destination: string;
  dates: string;
  description: string;
  coverImageUrl: string;
  placeCount: number;
  totalVotes: number;
  places: Place[];
}