# lib/api-zod/src/generated/types/tripSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TripSummary {
  id: number;
  title: string;
  destination: string;
  dates: string;
  description: string;
  coverImageUrl: string;
  placeCount: number;
  totalVotes: number;
  /** @nullable */
  leadingPlace: string | null;
}