# lib/api-zod/src/generated/types/tripVoteSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlaceVoteSummary } from './placeVoteSummary';

export interface TripVoteSummary {
  tripId: number;
  totalVotes: number;
  /** @nullable */
  leadingPlace: string | null;
  places: PlaceVoteSummary[];
}