# lib/api-zod/src/generated/types/voteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VoteInputMode } from './voteInputMode';

export interface VoteInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  voterId: string;
  mode: VoteInputMode;
}