# lib/api-zod/src/generated/types/voteResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VoteResultMode } from './voteResultMode';

export interface VoteResult {
  tripId: number;
  placeId: number;
  voteCount: number;
  totalVotes: number;
  isVoted: boolean;
  mode: VoteResultMode;
}