# lib/db/src/schema/trips.ts
import { createInsertSchema } from "drizzle-zod";
import { pgTable, serial, text } from "drizzle-orm/pg-core";
import { z } from "zod/v4";

export const tripsTable = pgTable("trips", {
  id: serial("id").primaryKey(),
  title: text("title").notNull(),
  destination: text("destination").notNull(),
  dates: text("dates").notNull(),
  description: text("description").notNull(),
  coverImageUrl: text("cover_image_url").notNull(),
});

export const insertTripSchema = createInsertSchema(tripsTable).omit({ id: true });
export type InsertTrip = z.infer<typeof insertTripSchema>;
export type Trip = typeof tripsTable.$inferSelect;